Add *Conn.CloseRead for WASM
